Various updates
1) Applied dodgy fix to make status on device card update (Line 375!),
without this line the status was always one change behind (I guess as
it was using oldStatus).

2) Added battery update function.

3) Added battery update to initialise function (Not sure if this is
best). Without it, on a clean install the device would not show a
battery reading until a change occurred.

3) Small format changes.

Thoughts, it seems the use of oldStatus / newStatus is starting to
become troublesome, I have a feeling we can refactor this and make it
much easier to maintain?